code shifted to [test_]image, files deleted
